+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [5.17.0] - 2026-03-02
 
 ### Added
+- Added support for insecure-skip-tls-verify during Kubernetes platform management
 - Added number of findings from the previous reporting period (`previous_period_findings`) for rules that are now deprecated in the OPERATIONAL_DEPRECATION report
 - Added Event-Driven (ED) jobs support for AMI installation
 - Added `STANDARD`, `SCHEDULED` to the `JobType` enum